FileAPI/url/url_xmlhttprequest.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>FileAPI Test: url_xmlhttprequest</title>
    <link rel="author" title="Intel" href="www.intel.com">
    <link rel="help" href="http://www.w3.org/TR/FileAPI/#creating-revoking">
    <link rel="help" href="http://www.w3.org/TR/XMLHttpRequest2/#interface-xmlhttprequest">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>

    <script>
      async_test(function () {
        var http = new XMLHttpRequest();
        http.open("GET", "test.png", true);
        http.responseType = "blob";
        http.onloadend = this.step_func(function (evt) {
          var blobURI = window.URL.createObjectURL(http.response);

          assert_true(http.response instanceof Blob, "XMLHttpRequest returns instanceof Blob");
          assert_equals(typeof blobURI, "string", "Blob URI is typeof string");
          assert_equals(blobURI.indexOf("blob"), 0, "Blob URI starts with 'blob'");
          assert_equals(http.status, 200, "The status is 200");
          assert_equals(http.statusText, "OK", "The status text is OK when XMLHttpRequest returns correct blob");
          assert_equals(http.getResponseHeader("Content-Type"), "image/png", "The content type is image/png when set the respnose blob");
          this.done();
        });
        http.send();
      });
    </script>
  </body>
</html>
